refactor explorer into seperate components
